refactor(perf): remove deprecated namespaced API
