+++ packages/app/lib/internal/constants.ts
@@ -31,7 +31,6 @@ export const KNOWN_NAMESPACES = [
   'storage',
   'naturalLanguage',
   'notifications',
-  'perf',
   'utils',
 ] as const;
