+++ packages/perf/e2e/perf.e2e.js
@@ -16,116 +16,6 @@
  */
 
 describe('perf() modular', function () {
-  describe('firebase v8 compatibility', function () {
-    beforeEach(async function beforeEachTest() {
-      // @ts-ignore
-      globalThis.RNFB_SILENCE_MODULAR_DEPRECATION_WARNINGS = true;
-    });
-
-    afterEach(async function afterEachTest() {
-      // @ts-ignore
-      globalThis.RNFB_SILENCE_MODULAR_DEPRECATION_WARNINGS = false;
-    });
-
-    describe('setPerformanceCollectionEnabled()', function () {
-      // These depend on `tests/firebase.json` having `perf_auto_collection_enabled` set to false the first time
-      // The setting is persisted across restarts, reset to false after for local runs where prefs are sticky
-      afterEach(async function () {
-        await firebase.perf().setPerformanceCollectionEnabled(false);
-      });
-
-      it('true', async function () {
-        should.equal(firebase.perf().isPerformanceCollectionEnabled, false);
-        await firebase.perf().setPerformanceCollectionEnabled(true);
-        should.equal(firebase.perf().isPerformanceCollectionEnabled, true);
-      });
-
-      it('false', async function () {
-        await firebase.perf().setPerformanceCollectionEnabled(false);
-        should.equal(firebase.perf().isPerformanceCollectionEnabled, false);
-      });
-    });
-
-    describe('instrumentationEnabled', function () {
-      afterEach(function () {
-        const perf = firebase.perf();
-        perf.instrumentationEnabled = false;
-      });
-
-      it('true', function () {
-        const perf = firebase.perf();
-
-        perf.instrumentationEnabled = true;
-
-        should.equal(perf.instrumentationEnabled, true);
-      });
-
-      it('should throw Error with wrong parameter', function () {
-        const perf = firebase.perf();
-
-        try {
-          perf.instrumentationEnabled = 'some string';
-
-          return Promise.reject(new Error('Did not throw Error.'));
-        } catch (e) {
-          e.message.should.containEql("'enabled' must be a boolean");
-          return Promise.resolve();
-        }
-      });
-    });
-
-    // TODO: flakey in Jet e2e.
-    xdescribe('dataCollectionEnabled', function () {
-      afterEach(function () {
-        const perf = firebase.perf();
-        perf.dataCollectionEnabled = false;
-      });
-
-      it('true', function () {
-        const perf = firebase.perf();
-
-        perf.dataCollectionEnabled = true;
-
-        should.equal(perf.dataCollectionEnabled, true);
-      });
-
-      it('should throw Error with wrong parameter', function () {
-        const perf = firebase.perf();
-
-        try {
-          perf.dataCollectionEnabled = 'some string';
-
-          return Promise.reject(new Error('Did not throw Error.'));
-        } catch (e) {
-          e.message.should.containEql("'enabled' must be a boolean");
-          return Promise.resolve();
-        }
-      });
-    });
-
-    describe('startTrace()', function () {
-      it('resolves a started instance of Trace', async function () {
-        const trace = await firebase.perf().startTrace('invertase');
-        trace.constructor.name.should.be.equal('Trace');
-        trace._identifier.should.equal('invertase');
-        trace._started.should.equal(true);
-        await trace.stop();
-      });
-    });
-
-    describe('startScreenTrace()', function () {
-      it('resolves a started instance of a ScreenTrace', async function () {
-        if (Platform.android) {
-          const screenTrace = await firebase.perf().startScreenTrace('FooScreen');
-          screenTrace.constructor.name.should.be.equal('ScreenTrace');
-          screenTrace._identifier.should.equal('FooScreen');
-          await screenTrace.stop();
-          screenTrace._stopped.should.equal(true);
-        }
-      });
-    });
-  });
-
   describe('modular', function () {
     describe('getPerformance', function () {
       it('pass app as argument', function () {
