+++ packages/perf/typedoc.json
@@ -1,5 +1,5 @@
 {
   "$schema": "https://typedoc.org/schema.json",
-  "entryPoints": ["lib/modular.ts", "lib/types/perf.ts"],
+  "entryPoints": ["lib/index.ts", "lib/types/perf.ts"],
   "tsconfig": "tsconfig.json"
 }
